+++ src/types/tasksLayoutType.ts
@@ -1,0 +1,5 @@
+// Tasks can be displayed in either a list or a grid format.
+
+type tasksLayoutType = "list" | "grid";
+
+export default tasksLayoutType;