+++ src/types/tasksViewType.ts
@@ -1,5 +1,0 @@
-// Tasks can be displayed in either a list or a grid format.
-
-type tasksViewType = "list" | "grid";
-
-export default tasksViewType;